Refactor redraw code so it never takes more than MAX_FRAME_TIME_MS.
Even "skip to end" redraws take a MAX_FRAME_TIME_MS and may return early.
This actually greatly simplifies the redraw logic, I think -- it eliminates
the distinction between "playback mode" and "refresh mode" that we had before.
One side-effect is that we can't safely use drawing.brush to mirror the
current UI brush state, since it's not guaranteed to be up to date at any
particular point.  So we split the brush_change commands up so include only
the brush information which is actually changing at a given point.