Add DejaVu Sans (Sugar's standard font) and use FontSquirrel to embed.
Used the settings in fonts/generator_config.txt to generated embedded
subsetted fonts from DejaVu Sans and Delius using:
  http://www.fontsquirrel.com/fontface/generator

We still need to use the webfont loaded to generate events, as the
fonts don't seem to be ready for use on <canvas> until the webfont loader
says they are.  Without the load event, the first letter you draw might
not have a "correct shape" overlay.