Store thumbnails as JPEGs, don't lose pixels on retina displays; fix shadows.
We store thumbnails as JPEGs to save some space. This requires us to also
make thumbnails opaque (since JPEG can't store transparency) which then
also requires a bit of CSS fixup to ensure that the inset shadows appear
over the (newly-opaque) thumbnails.

We also use the toDataURLHD method when saving layer checkpoints (including
thumbnails) to ensure that we don't lose resolution on retina displays.

Finally, only dump gallery thumbnails the first time the gallery is loaded,
so we don't waste too much storage on the logs.