Add fancier developer-save option to make it easier to export from Android.
Firefox/Android as a web-app doesn't seem to do anything useful with the
plain-text JSON download from the previous 'save' command.  You can't even
cut-and-paste it into an email.  Export as an encoded PNG instead, which
you should be able to save and send from Android.  Add a scripts to
uncompress and decode the resulting PNG as well.  Note that I attempted
to make the encoding well-behaved on retina devices, but I don't provide
any guarantees on the decoding end.