Some appearance work on the gallery; allow icons to be dragged.
The original goal here was to have the icons be dragged to the trashcan
in the toolbar to be deleted, but iframe issues make that difficult:
mousemove events stop as soon as the mouse moves over the toolbar
(although touchmove events continue just fine).  So we'll probably
revert some of this patch and implement some form of 'long tap to delete'
instead.