Drag and drop support for gallery thumbnails.
This was incredibly painful.  There are three different methods implemented:
 a) standard HTML5 Drag and Drop.  Works great, really easy to implement,
    doesn't work on mobile browsers (boo).
 b) touch events: touch occurs on element in inner iframe and touchmove
    events follow the gesture even when it leaves the iframe (good) but
    we need to element an overlay in the toolbar layer in order to get
    the drag shadow from being clipped in the iframe.
    but the drag shadow needs to be 'pointer-events: none' otherwise the
    touch events get interrupted.
 c) mouse events: mouse move events get interrupted when you hit the edge
    of the iframe (boo).  but we can get mouseover events (which don't
    exist for touch).  So as soon as we mouseover an element, move the
    drag shadow over it, so we can intercept events and handle the
    mousedown/mousemove directly on the drag shadow.

It all works (finally) but it's a bit ridiculous.  Probably method (c) could
have been eliminated, since browsers-with-mice tend to implement method (a)
(which is better, since it's not limited by the edges of the browser frame).
However, implementing (c) helped a lot with implementation and testing of (b),
which was the only one I *really* needed.  Sigh.